Fixing the use of the binary formatter in net461 so all unit tests now pass. Also removed the need to have it run only in framework as it also works on netcoreapp3.1
