+++ Octokit.Tests/Clients/UsersClientTests.cs
@@ -1,7 +1,4 @@
 ﻿using System;
-#if NET_45
-using System.Collections.ObjectModel;
-#endif
 using System.Threading.Tasks;
 using NSubstitute;
 using Octokit.Internal;
