+++ Octokit.Tests/Exceptions/ApiValidationExceptionTests.cs
@@ -1,12 +1,9 @@
 ﻿using System.Linq;
 using System.Net;
-#if !NO_SERIALIZABLE
-using System.Runtime.Serialization.Formatters.Binary;
-using System.IO;
-#endif
 using Xunit;
 
 using static Octokit.Internal.TestSetup;
+using Octokit.Tests.Helpers;
 
 namespace Octokit.Tests.Exceptions
 {
@@ -38,7 +35,6 @@ namespace Octokit.Tests.Exceptions
                 Assert.Equal("Validation Failed", exception.Message);
             }
 
-#if !NO_SERIALIZABLE
             [Fact]
             public void CanPopulateObjectFromSerializedData()
             {
@@ -48,18 +44,11 @@ namespace Octokit.Tests.Exceptions
                     @"already in use"",""resource"":""PublicKey""}],""message"":""Validation Failed""}");
 
                 var exception = new ApiValidationException(response);
+                var deserialized = BinaryFormatterExtensions.SerializeAndDeserializeObject(exception);
 
-                using (var stream = new MemoryStream())
-                {
-                    var formatter = new BinaryFormatter();
-                    formatter.Serialize(stream, exception);
-                    stream.Position = 0;
-                    var deserialized = (ApiValidationException)formatter.Deserialize(stream);
-                    Assert.Equal("Validation Failed", deserialized.ApiError.Message);
-                    Assert.Equal("key is already in use", exception.ApiError.Errors.First().Message);
-                }
+                Assert.Equal("Validation Failed", deserialized.ApiError.Message);
+                Assert.Equal("key is already in use", exception.ApiError.Errors.First().Message);
             }
-#endif
         }
     }
 }
